requirements.txt
flask==3.0.3
pillow==10.2.0
numpy==1.26.4
deepface==0.0.93
gunicorn==23.0.0
